Uploads name together with bot.

+++ Influence.Web/Controllers/UploadController.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.IO;
 using System.IO.Compression;
+using System.Text.RegularExpressions;
 using System.Web.Mvc;
 using Influence.Web.Models;
 using Exception = System.Exception;
@@ -11,32 +12,35 @@ namespace Influence.Web.Controllers
     {
         public static readonly string UploadedBotsFolderName = "UploadedBots";
 
-        public ActionResult Index()
-        {
-            return View(new UploadModel());
-        }
+        public ActionResult Index() 
+            => View(new UploadModel());
 
         [HttpPost]
         public ActionResult UploadFile(UploadModel model)
         {
             var file = model.File;
+            if (string.IsNullOrEmpty(model.Name))
+                return UploadResult("Fill out a name for your bot");
+            if (!Regex.IsMatch(model.Name, @"^[a-zA-Z0-9 æøåÆØÅ]*$"))
+                return UploadResult("The name can only contain letters, numbers and spaces");
             if (file?.ContentLength > 0 && !string.IsNullOrEmpty(file.FileName))
             {
                 try
                 {
-                    var fileName = Path.GetFileName(file.FileName);
+                    var zipFileName = Path.GetFileName(file.FileName);
 
                     if (Path.GetExtension(file.FileName) != ".zip")
                         return UploadResult("Uploaded file must be .zip");
 
                     var folderToPutFile = Server.MapPath($"~/{UploadedBotsFolderName}/{Guid.NewGuid()}");
                     var dirInfo = Directory.CreateDirectory(folderToPutFile);
-                    var fullPath = Path.Combine(dirInfo.FullName, fileName);
+                    var fullPath = Path.Combine(dirInfo.FullName, zipFileName);
                     file.SaveAs(fullPath);
 
                     ZipFile.ExtractToDirectory(fullPath, dirInfo.FullName);
                     System.IO.File.Delete(fullPath);
-
+                    string nameTxtPath = Path.Combine(folderToPutFile, "name.txt");
+                    System.IO.File.WriteAllText(nameTxtPath, model.Name);
                     return UploadResult("Success!");
                 }
                 catch (Exception ex)
@@ -48,7 +52,7 @@ namespace Influence.Web.Controllers
             return UploadResult("Choose a file");
         }
 
-        private ActionResult UploadResult(string message) 
+        private ActionResult UploadResult(string message)
             => View("Index", new UploadModel { Message = message });
     }
 }
